@@ -18,15 +18,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.3.2+3
-  barras: ^0.0.2
-  http: ^0.12.2
+  provider: 
+  # barras:
+  http:
+  firebase_core:
+  firebase_auth:
+  cloud_firestore: 
+
 
 
 
